+++ packages/cmd/src/syntax/SyntaxValidator.ts
@@ -294,8 +294,24 @@ export class SyntaxValidator {
 
                 // check if the value of the option conforms to the allowableValues (if any)
                 if (!isNullOrUndefined(optionDef.allowableValues)) {
-                    if (!this.checkIfAllowable(optionDef.allowableValues, commandArguments[optionName])) {
-                        this.invalidOptionError(optionDef, responseObject, commandArguments[optionName]);
+                    // Make a copy of optionDef, so that modifications below are only used in this place
+                    const optionDefCopy: ICommandOptionDefinition = JSON.parse(JSON.stringify(optionDef));
+                    // Use modified regular expressions for allowable values to check and to generate error
+                    optionDefCopy.allowableValues.values = optionDef.allowableValues.values.map((regex) => {
+                        // Prepend "^" if not existing
+                        if (!regex.startsWith("^")) {
+                            regex = "^" + regex;
+                        }
+
+                        // Append "$" if the last char is an escaped "$" or chars other than "$"
+                        if (!regex.endsWith("$") || regex.endsWith("\\$")) {
+                            regex = regex + "$";
+                        }
+
+                        return regex;
+                    });
+                    if (!this.checkIfAllowable(optionDefCopy.allowableValues, commandArguments[optionName])) {
+                        this.invalidOptionError(optionDefCopy, responseObject, commandArguments[optionName]);
                         valid = false;
                     }
                 }